feat : ajout des controllers de STORES

+++ stores/src/main/kotlin/iut/nantes/project/stores/Repository/StoreRepository.kt
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface StoreRepository : JpaRepository<StoreEntity, Long> {
+interface StoreRepository : JpaRepository<StoreEntity, Int> {
     fun findTopByOrderByIdDesc(): StoreDTO
 }
